Add files via upload
Payment Application that users have to input their PIN, Name, and Expiry date of the card. Terminal side checks whether this account is valid, expired, blocked or not, then the server side checks whether the amount of transaction is available for such an account or not, and if it exceeded the maximum transaction amount or not.